Lagt til enkle tester av nye utils-metoder og omstrukturert vilkårsvurderingDto-klasser
