+++ src/main/kotlin/no/nav/familie/ks/sak/api/dto/BehandlingResponsDto.kt
@@ -10,12 +10,6 @@ import no.nav.familie.ks.sak.kjerne.behandling.steg.BehandlingSteg
 import no.nav.familie.ks.sak.kjerne.behandling.steg.BehandlingStegStatus
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Målform
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonType
-import no.nav.familie.ks.sak.kjerne.vedtak.Standardbegrunnelse
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.AnnenVurderingType
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.Regelverk
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.Resultat
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.UtdypendeVilkårsvurdering
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.Vilkår
 import java.time.LocalDate
 import java.time.LocalDateTime
 
@@ -76,35 +70,4 @@ data class RegisteropplysningResponsDto(
     val fom: LocalDate?,
     val tom: LocalDate?,
     var verdi: String
-)
-
-data class PersonResultatResponsDto(
-    val personIdent: String,
-    val vilkårResultater: List<VilkårResultatDto>,
-    val andreVurderinger: List<AnnenVurderingDto>
-)
-
-data class VilkårResultatDto(
-    val id: Long,
-    val vilkårType: Vilkår,
-    val resultat: Resultat,
-    val periodeFom: LocalDate?,
-    val periodeTom: LocalDate?,
-    val begrunnelse: String,
-    val endretAv: String,
-    val endretTidspunkt: LocalDateTime,
-    val behandlingId: Long,
-    val erVurdert: Boolean = false,
-    val erAutomatiskVurdert: Boolean = false,
-    val erEksplisittAvslagPåSøknad: Boolean? = null,
-    val avslagBegrunnelser: List<Standardbegrunnelse>? = emptyList(),
-    val vurderesEtter: Regelverk? = null,
-    val utdypendeVilkårsvurderinger: List<UtdypendeVilkårsvurdering> = emptyList()
-)
-
-data class AnnenVurderingDto(
-    val id: Long,
-    val resultat: Resultat,
-    val type: AnnenVurderingType,
-    val begrunnelse: String?
 )
