+++ src/main/kotlin/no/nav/familie/ks/sak/api/mapper/BehandlingMapper.kt
@@ -1,20 +1,14 @@
 package no.nav.familie.ks.sak.api.mapper
 
 import no.nav.familie.kontrakter.felles.personopplysning.KJOENN
-import no.nav.familie.ks.sak.api.dto.AnnenVurderingDto
 import no.nav.familie.ks.sak.api.dto.ArbeidsfordelingResponsDto
 import no.nav.familie.ks.sak.api.dto.BehandlingResponsDto
 import no.nav.familie.ks.sak.api.dto.BehandlingStegTilstandResponsDto
 import no.nav.familie.ks.sak.api.dto.PersonResponsDto
-import no.nav.familie.ks.sak.api.dto.PersonResultatResponsDto
-import no.nav.familie.ks.sak.api.dto.VilkårResultatDto
 import no.nav.familie.ks.sak.kjerne.arbeidsfordeling.domene.ArbeidsfordelingPåBehandling
 import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.Person
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.AnnenVurdering
 import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.PersonResultat
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.Resultat
-import no.nav.familie.ks.sak.kjerne.vilkårsvurdering.domene.VilkårResultat
 
 object BehandlingMapper {
 
@@ -42,7 +36,8 @@ object BehandlingMapper {
             endretAv = behandling.endretAv,
             arbeidsfordelingPåBehandling = lagArbeidsfordelingRespons(arbeidsfordelingPåBehandling),
             personer = personer?.map { lagPersonRespons(it) } ?: emptyList(),
-            personResultater = personResultater?.map { lagPersonResultatRespons(it) } ?: emptyList()
+            personResultater = personResultater?.map { VilkårsvurderingMapper.lagPersonResultatRespons(it) }
+                ?: emptyList()
         )
 
     private fun lagArbeidsfordelingRespons(arbeidsfordelingPåBehandling: ArbeidsfordelingPåBehandling) =
@@ -60,36 +55,5 @@ object BehandlingMapper {
         kjønn = KJOENN.valueOf(person.kjønn.name),
         målform = person.målform,
         dødsfallDato = person.dødsfall?.dødsfallDato
-    )
-
-    private fun lagPersonResultatRespons(personResultat: PersonResultat) = PersonResultatResponsDto(
-        personIdent = personResultat.aktør.aktivFødselsnummer(),
-        vilkårResultater = personResultat.vilkårResultater.map { lagVilkårResultatRespons(it) },
-        andreVurderinger = personResultat.andreVurderinger.map { lagAnnenVurderingRespons(it) }
-    )
-
-    private fun lagVilkårResultatRespons(vilkårResultat: VilkårResultat) = VilkårResultatDto(
-        resultat = vilkårResultat.resultat,
-        erAutomatiskVurdert = vilkårResultat.erAutomatiskVurdert,
-        erEksplisittAvslagPåSøknad = vilkårResultat.erEksplisittAvslagPåSøknad,
-        id = vilkårResultat.id,
-        vilkårType = vilkårResultat.vilkårType,
-        periodeFom = vilkårResultat.periodeFom,
-        periodeTom = vilkårResultat.periodeTom,
-        begrunnelse = vilkårResultat.begrunnelse,
-        endretAv = vilkårResultat.endretAv,
-        endretTidspunkt = vilkårResultat.endretTidspunkt,
-        behandlingId = vilkårResultat.behandlingId,
-        erVurdert = vilkårResultat.resultat != Resultat.IKKE_VURDERT || vilkårResultat.versjon > 0,
-        avslagBegrunnelser = vilkårResultat.standardbegrunnelser,
-        vurderesEtter = vilkårResultat.vurderesEtter,
-        utdypendeVilkårsvurderinger = vilkårResultat.utdypendeVilkårsvurderinger
-    )
-
-    private fun lagAnnenVurderingRespons(annenVurdering: AnnenVurdering) = AnnenVurderingDto(
-        id = annenVurdering.id,
-        resultat = annenVurdering.resultat,
-        type = annenVurdering.type,
-        begrunnelse = annenVurdering.begrunnelse
     )
 }
